+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added weighting to text-mining analysis. Word clouds can now incorporate statistics.
 * Network plotting now performed using ggraph.
 * Removal of excessive warnings produced when performing text-mining.
+* Added PPI exploration of clusters
 
 # version 1.0.0